+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 ﻿# Change Log
 All notable changes to this project will be documented in this file.
 
-## [1.0.6] - 2016-11-11
+## [1.1.0] - 2016-11-14
+### Added
+- Added HTML tag `lang` attribute generation (based on "og:locale" setting).
+
+### Changed
+- `setTag` method overrides default settings.
+
 ### Fixed
+- Fixed "og:locale:alternate" issue, multiple tags are being generated for each available locale.
 - Fixed the "og:url" bug when applicationUrl is not supplied with the default settings.
 
 ## [1.0.5] - 2016-11-11